monster type refactor
