+++ src/Juno/Consensus/ByzRaft/Handler.hs
@@ -8,11 +8,11 @@ import Data.AffineSpace
 import Control.Monad
 import Data.Maybe
 
-import Juno.Runtime.Protocol.Types
-import Juno.Runtime.Log
+import Juno.Types
 import Juno.Consensus.ByzRaft.Commit (doCommit)
 import Juno.Runtime.Sender (sendAllAppendEntries,sendAllAppendEntriesResponse)
 import Juno.Util.Util (debug, dequeueEvent)
+
 import qualified Juno.Consensus.Pure.Handle.AppendEntries as PureAppendEntries
 import qualified Juno.Consensus.Pure.Handle.AppendEntriesResponse as PureAppendEntriesResponse
 import qualified Juno.Consensus.Pure.Handle.RequestVote as PureRequestVote
@@ -25,7 +25,7 @@ import qualified Juno.Consensus.Pure.Handle.Revolution as PureRevolution
 
 issueBatch :: Monad m => Raft m ()
 issueBatch = do
-  role' <- use role
+  role' <- use nodeRole
   ci <- use commitIndex
   case role' of
     Follower -> debug $ "Commit index is still: " ++ show ci
