+++ src/Juno/Consensus/ByzRaft/Server.hs
@@ -7,7 +7,7 @@ import qualified Data.Set as Set
 
 import Juno.Consensus.ByzRaft.Handler
 import Juno.Consensus.ByzRaft.Api (apiReceiver)
-import Juno.Runtime.Protocol.Types
+import Juno.Types
 import Juno.Util.Util
 import Juno.Runtime.Timer
 import Juno.Runtime.MessageReceiver
