+++ src/Juno/Consensus/Pure/Handle/HeartbeatTimeout.hs
@@ -15,10 +15,10 @@ import Juno.Consensus.Pure.Types
 import Juno.Runtime.Sender (sendAllAppendEntries)
 import Juno.Runtime.Timer (resetHeartbeatTimer, hasElectionTimerLeaderFired)
 import Juno.Util.Util (debug,enqueueEvent)
-import qualified Juno.Runtime.Protocol.Types as JT
+import qualified Juno.Types as JT
 
 data HeartbeatTimeoutEnv = HeartbeatTimeoutEnv {
-      _role :: Role
+      _nodeRole :: Role
     , _leaderWithoutFollowers :: Bool
 }
 makeLenses ''HeartbeatTimeoutEnv
@@ -28,13 +28,13 @@ data HeartbeatTimeoutOut = IsLeader | NotLeader | NoFollowers
 handleHeartbeatTimeout :: (MonadReader HeartbeatTimeoutEnv m, MonadWriter [String] m) => String -> m HeartbeatTimeoutOut
 handleHeartbeatTimeout s = do
   tell ["heartbeat timeout: " ++ s]
-  role' <- view role
+  role' <- view nodeRole
   leaderWithoutFollowers' <- view leaderWithoutFollowers
   case role' of
     Leader -> if leaderWithoutFollowers'
               then tell ["Leader found to not have followers"] >> return NoFollowers
-              else return $ IsLeader
-    _ -> return $ NotLeader
+              else return IsLeader
+    _ -> return NotLeader
 
 handle :: Monad m => String -> JT.Raft m ()
 handle msg = do
@@ -42,7 +42,7 @@ handle msg = do
   leaderWithoutFollowers' <- hasElectionTimerLeaderFired
   (out,l) <- runReaderT (runWriterT (handleHeartbeatTimeout msg)) $
              HeartbeatTimeoutEnv
-             (JT._role s)
+             (JT._nodeRole s)
              leaderWithoutFollowers'
   mapM_ debug l
   case out of
@@ -53,5 +53,5 @@ handle msg = do
       JT.timeSinceLastAER %= (+ hbMicrosecs)
     NotLeader -> JT.timeSinceLastAER .= 0 -- probably overkill, but nice to know this gets set to 0 if not leader
     NoFollowers -> do
-      timeout' <- return $ (JT._timeSinceLastAER s)
-      enqueueEvent $ (ElectionTimeout $ "Leader has not hear from followers in: " ++ show (timeout' `div` 1000) ++ "ms")
+      timeout' <- return $ JT._timeSinceLastAER s
+      enqueueEvent $ ElectionTimeout $ "Leader has not hear from followers in: " ++ show (timeout' `div` 1000) ++ "ms"
