+++ src/Juno/Consensus/Pure/Types.hs
@@ -32,5 +32,4 @@ module Juno.Consensus.Pure.Types (
   ) where
 
 
-import Juno.Runtime.Protocol.Types
-import Juno.Runtime.Types
+import Juno.Types
