+++ src/Juno/Messaging/ZMQ.hs
@@ -19,8 +19,7 @@ import Data.Thyme.Clock
 import Data.Serialize
 
 import Juno.Messaging.Types
-import Juno.Runtime.Types (ReceivedAt(..),Digest(..),MsgType(..))
-import Juno.Runtime.Protocol.Types (SignedRPC(..))
+import Juno.Types (ReceivedAt(..),Digest(..),MsgType(..), SignedRPC(..))
 
 sendProcess :: OutChan (OutBoundMsg String ByteString)
             -> Rolodex String (Socket z Push)
