+++ src/Juno/Runtime/MessageReceiver.hs
@@ -15,8 +15,7 @@ import qualified Data.Map.Strict as Map
 import qualified Data.Serialize as S
 import qualified Data.Set as Set
 
-import Juno.Runtime.Types
-import Juno.Runtime.Protocol.Types
+import Juno.Types
 import Juno.Util.Util (debugNoRole,enqueueEvent)
