+++ src/Juno/Spec/Simple.hs
@@ -12,9 +12,8 @@ module Juno.Spec.Simple
 
 import Juno.Consensus.ByzRaft.Server
 import Juno.Consensus.ByzRaft.Client
-import Juno.Runtime.Types
 import Juno.Messaging.Types
-import Juno.Runtime.Protocol.Types
+import Juno.Types
 import Juno.Messaging.ZMQ
 import Juno.Monitoring.Server (startMonitoring)
 import Juno.Runtime.Api.ApiServer
